EIP1-14024: refactor: - extract backup logic to functions file to be testable - stricter typing - update node version to reduce warnings in workflow output
