+++ .github/workflows/backup-notify-templates.yml
@@ -12,9 +12,9 @@ jobs:
   backup-templates:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
 
-      - uses: actions/setup-node@v4
+      - uses: actions/setup-node@v5
         with:
           node-version: 20
