+++ templates/package.json
@@ -6,6 +6,7 @@
     "backup": "tsx backup-templates.ts"
   },
   "devDependencies": {
+    "@types/node": "^20.14.0",
     "tsx": "^4.19.0",
     "typescript": "^5.5.0"
   },
